Properly nest all headings for correct bookmarks index
All sections and articles would start with a <h1> which would mess up
the autogenerated bookmarks index displayed by the PDF viewer.

Initially I experimented with increasing the heading levels of a page
based on how deeply nested under subsections the page is.
This had some issues:

1. If a page was nested under N sections/dirs and used headings up to
   level L where N + L > 6, the headings' level couldn't be increased
   any more. I resorted to transforming them to `<strong>`.
1. Pages were now themed differently. The title of a page would be an
   `<h1>` if on top level, `<h2>` if under a top level section, `<h3>`
   if under a subsection and so on.
1. The top page banner that indicates the current document would just
   display the top level section for non top-level pages.

In the end I decided that it's best to not mess with the heading levels
of pages, just with their nesting in the PDF's bookmark index.

This was achieved via the use of the `bookmark-level` CSS property.

Therefore this commit shouldn't affect the contents of the output PDF at
all, just the nesting of the pages & sections in its bookmark index.