Remove page edit & other links added by material theme
For example the page edit link (pencil icon) and any other links added
by material theme on the right of the page header, inside the article
content.

Without this change, there'd be no icon but the entire URL for editing
the page would be included in the PDF above each page title which was
undesired.

Filtering of such links is based on the `md-content__button` class as
suggested in https://github.com/squidfunk/mkdocs-material/issues/1920